+++ app/src/main/java/com/example/lab2/MainActivity.kt
@@ -9,6 +9,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.lab2.ui.CanvasScreen
+import com.example.lab2.ui.DrawingListScreen
+import com.example.lab2.ui.StartScreen
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -102,15 +105,23 @@ class MainActivity : AppCompatActivity() {
 fun ComposeNavigation() {
     val navController = rememberNavController()
 
-    NavHost(navController = navController, startDestination = "StartScreen") {
-        composable("StartScreen") {
-            StartScreen(onNavigate = { drawingId ->
-                navController.navigate("CanvasScreen/$drawingId")
-            })
-        }
-        composable("CanvasScreen/{drawingId}") { backStackEntry ->
-            val drawingId = backStackEntry.arguments?.getString("drawingId")?.toIntOrNull() ?: 0
-            CanvasScreen(navController = navController, drawingId = drawingId)
+//    NavHost(navController = navController, startDestination = "StartScreen") {
+//        composable("StartScreen") {
+//            StartScreen(onNavigate = { drawingId ->
+//                navController.navigate("CanvasScreen/$drawingId")
+//            })
+//        }
+//        composable("CanvasScreen/{drawingId}") { backStackEntry ->
+//            val drawingId = backStackEntry.arguments?.getString("drawingId")?.toIntOrNull() ?: 0
+//            CanvasScreen(navController = navController, drawingId = drawingId)
+//        }
+//    }
+    NavHost(navController = navController, startDestination = "start_screen") {
+        composable("start_screen") { StartScreen(navController) }
+        composable("canvas_screen") { CanvasScreen(navController = navController, drawingId = null) } // 这里处理新画布
+        composable("canvas_screen/{drawingId}") { backStackEntry ->
+            val drawingId = backStackEntry.arguments?.getString("drawingId")?.toInt()
+            CanvasScreen(navController, drawingId) // 传递drawingId
         }
     }
 }
