--- app/src/main/java/com/example/lab2/StartScreen.kt
@@ -1,16 +1,0 @@
-//package com.example.lab2.ui.screens
-//
-//import androidx.compose.runtime.Composable
-//import androidx.compose.runtime.collectAsState
-//import androidx.compose.ui.Alignment
-//import androidx.compose.ui.Modifier
-//import androidx.compose.material3.Button
-//import androidx.compose.material3.Text
-//import android.widget.Toast
-//import androidx.compose.ui.platform.LocalContext
-//import androidx.compose.foundation.layout.Box
-//import androidx.compose.foundation.layout.fillMaxSize
-//import androidx.compose.runtime.getValue
-//import androidx.lifecycle.viewmodel.compose.viewModel
-//import com.example.lab2.CustomViewModel
-//
